misc/ios: support go run in go_darwin_arm_exec
The -lldb flag makes it easy to use go run and end up in a debugging
session on darwin/arm.

Also: modify what text we look for on the getwd breakpoint. About
one time in a hundred we do not see "stop reason = breakpoint", which
is causing flakiness on the builder. The "->" marker appears more
often, so giving that a try.

Change-Id: I556f93e950086a7dff4839f301b9c55f7579f87b