runtime: provide an entry point for external code to initialize the runtime (on amd64).
The main purpose is to allow external code that wants to use a Go shared library
(i.e., Go code compiled with the "-shared" flag) to initialize the runtime.
(The external code still needs to pass in the correct argc and argv struct
parameters to the Go runtime.)

This change follows the changes already made for android/arm.

Change-Id: I81d75bf5d9b7c00c622fd6b975a6401d365c9166