Update patch set 1
Patch Set 1:

filepath.Join also have this behavior. Rather than
duplicated the docs for Join (or make the reader
for Join also read the docs for Clean), I'm wonder
what if we should still write this at the package
level, and note exceptions in the docs for Split.

This is because it's normal for users to assume this
behavior for all functions in this package, so we
make it clear.

What do you think?

Patch-set: 1