Update patch set 1
Patch Set 1: Code-Review-1

As discussed in the execution modes doc, I would like it to be possible to create a shared library written in Go that can be used as a plugin for a C program.  That implies that the shared library will initialize itself; no initialization routine will be called.  So I think we should implement that, in which case I think this change is unnecessary.

If we generate the shared library using internal linking, we can set DT_INIT to whatever we like.  If we generate the shared library with external linking, we can add an entry to a .init_array section and rely on the external linker.

Patch-set: 1
Reviewer: Gerrit User 5206 <5206@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=-1