runtime: Remove write barriers during STW.
When the GC does a stop the world code still needs to be run after the
GC believes the world is stopped.
One example is code that enqueues stopped gs onto runnable queues.
Another is when routine return from syscalls.
When this was done in C it was simple enough to avoid write barriers
but now that the GC and scheduler is written in Go these write barriers
need to be avoided. Since we don't move objects the only requirement is that
at least one of the paths to a live object is followed. One particularly
sticky problem was if a write barrier happened while the GC thought the
mutator was stopped an object might be marked but the write barrier might be
delayed in informing the GC to scan an object. Since other
paths to the object will see the object as marked they will not
inform the GC that the object needs to be scanned. The GC will move on
and the object will be incorrectly reclaimed. It is never the case that
these writes own the only path to the schedulers objects.

Some of the trace code was turned off, in particular the code that traces
returning from a syscall. The GC assumes that as far as the heap is concerned
the thread is stopped when it is in a syscall. Upon returning the trace code
must not do any heap writes for the same reasons discussed above.

Change-Id: Ic2e70b7caffa053e56156838eb8d89503e3c0c8a