Update patch set 2
Patch Set 2:

Ummm... this is lots of clutter throughout the runtime. I suspect there are more places that do pointer stores without a P, the two places I can think of are injectglist and mutexes (which build M list in semaphore implementation). We also need to move timers into netpoll, that is, they will run without a P.

What's wrong with traceGoSysExit? What write barrier does it do? Do we have a plan for fixing it?

If the problem is only with m/p/g, then we can move them off heap and consider as roots.

However that won't work for timers (they need to manipulate heap of heap-allocated object) and maybe for traceGoSysExit. So instead we can just fix the GC to not throw pointers on the floor. Threads that run without a P get workbufs from a global list, and this is a natural point for synchronization with these threads. You just need to ensure that a thread grabs workbuf, marks an object and puts workbuf, all atomically with respect to GC. This looks like the right solution to me.

Patch-set: 2