Update patch set 2
Patch Set 2:

I'll look at the other places you mention, thanks.

As you note write barriers must be atomic w.r.t. to the GC. Consider what can happen if they aren't.

We have an object x that holds a pointer to y and a reference to x is written into a marked object.

a. The write barrier marks object x as reachable and then is interrupted, say by being preempted by the OS, before it can put x into a wbuf. A marked object not in a wbuf is considered black.
b. The GC, thinking the world is stopped scans all objects in the wbuf which blackens all grey object.
c. The GC, thinking there are no grey object left concludes the mark phase and does a sweep. Since x was not scanned y never gets marked and the sweep finds an unmarked object y and frees it.
d. Sometime later the mutator in step a is resumed and writes x into a wbuf assuming that the GC will scan x and find y. Unfortunately the GC has already freed y.

Stop the world means the world is stopped and there is no heap mutation. If a mutator is going to do something when it says it is stopped it needs to make sure the GC fully understands what it is doing. This CL adds a check in the write barrier that throws when it encounter these situations.

The problem with the trace code is that it is very general and runs at times when the GC thinks the mutator is stopped. It flushes buffers when they are full using straight forward techniques that include writing pointers. 

We (I) do not have a well thought out plan/design for the trace routines. My preference is to not allow tracing when the GC thinks the world is stopped. In the case of exiting a syscall I just removed that event hoping that whatever event happens next implies that the syscall has exited. If other events can happen during a STW we can use the same or a similar approach. Again this is just an idea, not a plan/design tracer proposal.

Patch-set: 2
Label: Code-Review=0