Update patch set 2
Patch Set 2:

> > Your patch is relying on the fact that for a symbol Sym GCC will
 > > emit __imp__Sym.
 > 
 > I don't see how you make that conclusion. My changes do not affect
 > symbols starting with __imp__. I don't even consider them.

In ldpe.go, it seems to me that the code used to change __imp_Foo to Foo.  With this patch, it no longer does that.  I don't understand how that will interact with Windows/amd64, on which native symbols do not have a leading underscore.  Perhaps I misunderstand, but that is what I am talking about.


 > What I observe is that DLL function (like GetCurrentThreadId)
 > references are sometimes compiled as GetCurrentThreadId, and
 > sometimes compiled as __imp_GetCurrentThreadId by gcc. Our current
 > ldpe.go code threat them as the same symbol (it ignores / removes
 > __imp_ prefix). And that is invalid, because gcc generates
 > different code to access both.
 > 
 > My change stops them been bundled together by removing the code
 > that ignores / remove __imp_ prefix. So now both GetCurrentThreadId
 > and __imp_GetCurrentThreadId symbols stay separate. I also add
 > extra line "//go:cgo_import_dynamic __imp_GetCurrentThreadId ..."
 > in cgo, otherwise linker complains that __imp_GetCurrentThreadId is
 > not defined.

Right, but as far as I can tell this only works if all user defined symbols wind up with a leading underscore, which is not true on Windows/amd64.

Patch-set: 2