Update patch set 1
Patch Set 1:

I suppose my concern is that your step 4 requires knowing something special about the runtime and crypto/x509 packages.  You know that, but effectively nobody else does.  So this will become a bit of functionality in the test that nobody really uses.

If this could be part of some larger scheme, ideally one that can be automatically be invoked by run.bash in some documented manner, then perhaps I could see it.  But as a standalone change I don't find it convincing.

Patch-set: 1