Update patch set 1
Patch Set 1:

> I suppose my concern is that your step 4 requires knowing something
 > special about the runtime and crypto/x509 packages.  You know that,
 > but effectively nobody else does.  So this will become a bit of
 > functionality in the test that nobody really uses.
 > 
 > If this could be part of some larger scheme, ideally one that can
 > be automatically be invoked by run.bash in some documented manner,
 > then perhaps I could see it.  But as a standalone change I don't
 > find it convincing.

One option is to print an informative message when skipping a test in
step 3. Something like:
"Skipping because 'go' command is not available in $PATH.
Re-run the test locally with -target flag."

But this still relies on the user to see the message and act on it.

A better, and more defined way would be to add a -target flag to run.bash.
Under the flag, all test compilation/execution steps should happen through
go_target and go_target_exec wrappers. run.bash (now "go tool dist test")
will itself use the wrappers to run "go test", and when necessary, pass the
flag to the tests themselves, so they can in turn compile/run for the target.

In all my remote testing, I've been using my own version of run.bash tuned
for this purpose. For example, to run the standard library tests for -target
"daisy", it does the two step process in the CL description:
$ go_daisy test -p=1 -exec="go_daisy_exec" std
$ go test -p=1 runtime crypto/x509 -target="daisy"

Also, with -target support in run.bash, there is no need for the automatic
go_${GOOS}_${GOARCH}_exec lookup mechanism in go "run"/"test" commands that
"nacl" and "android" rely on. It can probably be made redundant and removed
(explicit is better than implicit).

Let me know if you'd like me to send a CL with -target support in run.bash.
It would probably touch a handful of modules: run.bash, cmd/dist, test,
with runtime, crypto/x509 changes already in this CL.

Patch-set: 1