Update patch set 2
Patch Set 2:

> What do you think about making GC synchronize with writebarriers
 > running w/o P (the idea I outlined in the comment above)? It will
 > not require any changes to the bulk of the runtime code and will
 > allow tracing and netpoll and timers without P.

I am very strongly in favor of the "no P = no write barriers or blow up" rule. Stoptheworld is about stopping the P's. That's our working definition of "the world". If you don't have a P, you're not part of "the world" and you don't get to do operations on the mutator graph. It's simple, it's easy to understand, and it's easy to verify at run time.

If we start making exceptions for write barriers when there is no P, the whole thing gets much more complex, much harder to understand, and much harder to verify. There would need to be a VERY strong argument for why that complexity is needed, why we'd want to maintain and pay the cost of that complexity for the forseeable future, and I just don't see that argument here.

For the specific case of tracing when exiting a system call, if you want to restore that record, you are welcome to, but you must do it without heap mutations. One possibility would be to delay the emitting of the record until the goroutine has reacquired a P. Or perhaps when the goroutine starts running again, the record that it has been scheduled could include an additional time stamp saying when it came out of the system call (instead of that information being its own record).

Patch-set: 2