Update patch set 2
Patch Set 2:

> For the specific case of tracing when exiting a system call, if you
 > want to restore that record, you are welcome to, but you must do it
 > without heap mutations. One possibility would be to delay the
 > emitting of the record until the goroutine has reacquired a P. Or
 > perhaps when the goroutine starts running again, the record that it
 > has been scheduled could include an additional time stamp saying
 > when it came out of the system call (instead of that information
 > being its own record).

For the tracer, it's possible the only time it modifies the mutator graph is when it fills up a trace buffer. That suggests another solution to this part of the problem might be to ensure there's always enough buffer slack when the world is stopped to absorb these events. This could be some allocation that happens when stopping the world, or perhaps each M could have enough buffer for one pending trace record that gets used when there's no P and flushed to the real trace buffer when the M starts again.

Patch-set: 2