Update patch set 2
Patch Set 2:

If you compile "void Foo() { }" and then run nm on the resulting object file, on windows/386 you will see "_Foo" and on windows/amd64 you will see "Foo".  In other words, from the object file perspective, which is what the linker works with, on windows/386 all user defined symbols have a leading underscore, and on windows/amd64 they do not.

Perhaps our ongoing confusion is because you are speaking from the perspective of the programmer and I am speaking from the perspective of the linker.  Or perhaps not, I'm not sure.

Patch-set: 2