Update patch set 3
Patch Set 3:

> > Can you do this for the copy of this code in ld too? (Or remove
 > the
 > > duplication somehow?)
 > 
 > I'm pretty sure I already did it for that copy.
 > 
 > The duplication is intentional. It will be easier to streamline the
 > two versions if the two uses (generating individual object files vs
 > linking many) are kept separate.

Oh so you did, sorry about that.  If the duplication is intentional, then you'd consider patches removing fields from LSym that are not used in internal/obj's version?

It seems to me that only having one definition of R_CALL and other constants that are written to object files would be a good thing though?

Patch-set: 3