Update patch set 3
Patch Set 3: -Code-Review

> Section 9.7 of the requirement states:

Ok, that is a good reason to include them.

However, we don't have to worry about the DirectoryNames, right? I could be convinced if CAs are actually including DirectoryName in real certificates but the example (which covers pretty much the only case for this) doesn't have them.

Also, I don't believe that we can break the API, even with a cmd/fix. I think the critical flag need to keep its (slightly anachronistic) name.

Lastly, if we're going to parse these then Verify will have to enforce them. (Although that doesn't have to be in the same change.)

Patch-set: 3
Label: Code-Review=0