Update patch set 2
Patch Set 2:

> 
 > Your change should work fine if all symbols start with an
 > underscore.  I'm worried about the case where symbols do not start
 > with an underscore, for which your change to the "__imp_" handling
 > will change behaviour.
 > 

Can you explain which part of my CL (or any explanation of mine) will not work if symbols do not start with an underscore. Perhaps an example to demonstrate your idea.

If gcc gives us __imp_F, and ldpe renames it to F, then surely we should have some test that will break, if we stop renaming (like this CL does). None of our tests breaks.

Alex

Patch-set: 2