net/http: Allow double-quoted cookie values
RFC 6265[1] allows cookie values to be double-quoted[2].  However, the
current sanitization code strips the double quotes out of a double-quoted
value (i.e., a value beginning with a double quote and ending with a double
quote).  Ironically, the sanitization code double quotes a value if it begins
or ends with a space or comma.  The RFC grammar specification allowing the
surrounding double quotes is also included in the comments of the
sanitizeCookieValue() function.

This changeset will allow a user to set a double-quoted cookie value.

[1] http://tools.ietf.org/html/rfc6265
[2] cookie-value = *cookie-octet / ( DQUOTE *cookie-octet DQUOTE )

Change-Id: Id87b1a1199b699f5cd24b4e9cd30a3c16d42098b