Update patch set 3
Patch Set 3:

DirectoryNames are used to restrict the values of the subject of endentity certificates and are therefore an important part of the Name Constraints.

I'm not that sure of the cmd/fix either, it's also not fixing all references as I highlighted om the golang-nuts group. While it feels completely wrong to keep the name PermittedDNSDomainsCritical as it's marking the extension as critical but is or wasn't doing anything with DNSNames itself.

Is there a preferred way to implement these API changes? 

I can imagine that we split this change in two one that adds the values and a second one that renames the critical flag for a next major release of GO?

Another option would be to map PermittedDNSDomainsCritical to NameConstraintsCritical with a deprecated comment in the documentation and remove the old one in a next major version?

The next step would be to actually implement name constraint verification in x509.Certificate.Verify(). But I have limited time to work on this, so I would like to do it step by step.

Patch-set: 3
Label: Code-Review=0