Update patch set 1
Patch Set 1:

This removes the alloc from the string concatenation, in a simple and re-usable way. A quick benchmark shows it being faster than concatenation as well.

I tried a dedicated conv string-building function, but it was bit ugly, since it had to handle convX2X, assertX2X, and assertX2X2. This seemed nicer.

And I got nervous about caching the result from syslook. substArgTypes modifies the node, so I'm not sure that it is safe to re-use them.

Patch-set: 1
Label: Code-Review=0