Update patch set 4
Patch Set 4:

I agree. Please wait for more review on changes like these. I will revert this one. If the computation of convE2T etc is causing problems, it is trivial to use a switch or an index into a list since there are only a few cases.

Generally, unless you can show a direct, significant performance impact, it's not worth making the code more complex.

For the specific case of right now, I am still trying to get to the bottom of the performance difference between when this code was C compiled with gcc and now (Go compiled with itself). It would help not to have performance tweaks moving the goalposts.

Patch-set: 4
Reviewer: Gerrit User 5056 <5056@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=0