encoding/json: Optionally preserve the key order of JSON objects
By default, JSON objects are decoded into an interface{} as a
map[string]interface{}, which discards the order of keys in the original object,
due to the randomized order of Go maps.

This CL implements an optional OrderedObject type as well as UseOrderedType on
the JSON Decoder to allow preservation of key order. This has no effect on
existing code. It is opt-in only.

This CL also supports encoding JSON objects in a determined order using the
OrderedObject type.

Change-Id: Ic325b2f8207408c00b38e53998c1d22227140113