Update patch set 3
Patch Set 3:

Just from reading the CL description, it's not clear to me there is actually any problem to fix. That said your description is confusing and vague, so I may be missing the point.

When typing at a Unix terminal, ^D flushes the input. That's all it does. If there is no text on the line, this results in a 0-length read, which most tools decide treat as EOF. But it is not strictly EOF; one can keep reading and see the next line of input.

Thus the scanner is behaving normally, as I read your description. It seems fine that after a pseudo-EOF caused by ^D, one can keep reading the input if one so chooses.

I believe what you're trying to achieve here is a sticky EOF, but that is a change of behavior I do not think is correct or warranted.

Also I have no idea what BOMs have to do with anything here.

Patch-set: 3
Reviewer: Gerrit User 5015 <5015@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=0