Update patch set 3
Patch Set 3:

Hi Rob, sorry that I didn't make myself clear. Another way to put it is that Peek() uses scanner.EOF to decide that no data has been scanned, resulting in it trying to read past a detected EOF.

I agree with you that, if the user wishes to read more data ofter getting an EOF, that's fine and thats what Next() can do. But the bug is that Peek() actually reads data itself (rather than just peeking) if it thinks no data has been scanned yet. It does this because it tries to detect and discard a BOM marker.

Does that make more sense?

Patch-set: 3