Update patch set 3
Patch Set 3:

One more quick followup for explanation.

I was wrong about the reason for Peek() reading data itself. I indicated it was to check for the BOM marker at the beginning of the file.

Really, it's that Peek() attempts to read in a token if neither Next() nor Scan() have been called yet. So Peek() isn't strictly peeking at the next token, it might go out and get that next token if it detects that it's companion functions haven't populated the current token yet.

Thusly, the bug is that Peek() mistakenly believes that s.ch has not yet been populated when Next() or Scan() set s.ch to EOF.

Patch-set: 3