Update patch set 3
Patch Set 3:

> You want rather import "image/internal/imageinternal"? That stutter seems like overkill here.

or image/internal/imgutil or /draw or something.

> Or are you saying that the package should be "package image" or "package imageinternal" even
> though the path is "image/internal", so that path basename != package name?

Oh, no, definitely not that.

We can just rename on import if it gets to that. I do expect use of internal across the board to grow over time, so I do think that's a likely outcome. But we can also just rename stuff later, too.

Patch-set: 3