Update patch set 5
Patch Set 5:

We are not supposed to be adding stuff to syscall any more unless it is required to support the core libraries. I need convincing this meets the bar.

In any case, this belongs in x/sys/unix as well as or instead of here.

Patch-set: 5
Reviewer: Gerrit User 5015 <5015@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=0