Update patch set 5
Patch Set 5:

> We are not supposed to be adding stuff to syscall any more unless
 > it is required to support the core libraries. I need convincing
 > this meets the bar.

This kind of change must go in the standard library.  It's a change to the fork/exec support in the os/exec package.  There is no way to add to that support except by changing the SysProcAttr in the standard syscall package.

 > In any case, this belongs in x/sys/unix as well as or instead of
 > here.

The files being changed by this CL do not exist in x/sys/unix.  There is no need for them there, because these files only exist to support the os/exec package, and are not called by anything else.

Patch-set: 5