Update patch set 5
Patch Set 5:

Thanks for the info, some of which belongs in the CL description (I'm starting to sound like a broken record on that point, but I won't stop.)

Please do not check stuff into syscall without my approval. I'm in the process of pulling x/sys back together and want to control the divergence.

Patch-set: 5