Update patch set 5
Patch Set 5:

> Thanks for the info, some of which belongs in the CL description
 > (I'm starting to sound like a broken record on that point, but I
 > won't stop.)
 > 
 > Please do not check stuff into syscall without my approval. I'm in
 > the process of pulling x/sys back together and want to control the
 > divergence.

My apologies for not waiting for approval.  This is actually a reprise of a CL that was approved and committed before (http://golang.org/cl/131750044), and was rolled back during the 1.4 release process because it introduced a bug.  But that should have been in the CL description, and was not.

Patch-set: 5