Update patch set 2
Patch Set 2:

Russ,
this change is different from the rest of the escape changes. It extends the existing escape analysis algo to a new type of objects. It is similar to the other changes I submitted earlier for strings and maps. You said that such changes can go in, but just wait for C->Go conversion.
Did you abandon it by error, or something has changed?

Patch-set: 2