Update patch set 1
Patch Set 1:

Looks like this doesn't work. The builder is reporting plenty of:

   error: Process is running.  Use 'process interrupt' to pause execution.

which means that some backtraces are being printed before the breakpoint has hit.

Tomorrow I'll change it to go back to waiting for "reason = breakpoint", but insert a 1s timeout to push ahead. It will make the code a bit messier, but will hopefully catch this occasional stuck state.

Patch-set: 1