cmd/internal/gc: improve flow of input params to output params
This includes the following information in the per-function summary:

outK = paramJ   encoded in outK bits for paramJ
outK = *paramJ  encoded in outK bits for paramJ
heap = paramJ   EscHeap
heap = *paramJ  EscContentEscapes

Note that (currently) if the address of a parameter is taken and
returned, necessarily a heap allocation occurred to contain that
reference, and the heap can never refer to stack, therefore the
parameter and everything downstream from it escapes to the heap.

The per-function summary information now has a tuneable number of bits
(2 is probably noticeably better than 1, 3 is likely overkill, but it
is now easy to check and the -m debugging output includes information
that allows you to figure out if more would be better.)

A new test was  added to check pointer flow through struct-typed and
*struct-typed parameters and returns; some of these are sensitive to
the number of summary bits, and ought to yield better results with a
more competent escape analysis algorithm.  Another new test checks
(some) correctness with array parameters, results, and operations.

The old analysis inferred a piece of plan9 runtime was non-escaping by
counteracting overconservative analysis with buggy analysis; with the
bug fixed, the result was too conservative (and it's not easy to fix
in this framework) so the source code was tweaked to get the desired
result.  A test was added against the discovered bug.

The escape analysis was further improved by splitting the "level" into
a negative-positive pair; normally the level is kept in a canonical
form where the sign indicates which field of the pair contains the
value (they can always be combined by adding) but certain dereference
operations increment the positive field even if the negative field is
not zero; a subsequent address-of (which decrements) checks to see if
it is paired with a dereference, and in that case they cancel (and the
addressed thing is not exposed to the heap).

This has been benchmarked (using the GC in its current under-
development state, so these results are not entirely reliable) using
600 repetitions of a compiler run with tight GC constraints and
counting # of collections:

cd html/template
for i in {1..600} ; do
   GODEBUG="gctrace=1" GOGC=10 go tool 6g *.go >& tmp.log
   tail -1 < tmp.log >> master_summary.log
done

with these results for # of GCs -- about 6% fewer.

35.5   33.1  average
36     33    median
47     40    max
25     24    min
2.97    2.46 "stdev"
master test

Number of GCs
is the metric used because better escape analysis should only effect
how often the GC is run -- anything that can be inferred not to escape
is generally short-lived and soon dead and won't change the size of
"live" much.

Update #3753
Update #4720
Fixes #10466

Change-Id: I0fd97d5f5ac527b45f49e2218d158a6e89951432