cmd/internal/gc: improve flow of input params to output params
This includes the following information in the per-function summary:

outK = paramJ   encoded in outK bits for paramJ
outK = *paramJ  encoded in outK bits for paramJ
heap = paramJ   EscHeap
heap = *paramJ  EscContentEscapes

Note that (currently) if the address of a parameter is taken and returned,
necessarily a heap allocation occurred to contain that reference, and the
heap can never refer to stack, therefore the parameter and everything
downstream from it escapes to the heap.

The per-function summary information now has a tuneable number of bits
(2 is probably noticeably better than 1, 3 is likely overkill,
but it is now easy to check and the -m debugging output includes information
that allows you to figure out if more would be better.)

New tests were added to check pointer flow through struct-typed and *struct-typed
parameters and returns; some of these are sensitive to the number of summary bits,
and ought to yield better results with a more competent escape analysis algorithm.

The old analysis inferred a piece of plan9 runtime was non-escaping by counteracting
overconservative analysis with buggy analysis; with the bug fixed, the result
was too conservative (and it's not easy to fix in this framework) so the source
code was tweaked to get the desired result.  A test was added against the
discovered bug.

Change-Id: I0fd97d5f5ac527b45f49e2218d158a6e89951432