cmd/internal/obj, etc: replace -shared with -buildmode=c-shared
Currently some compilers and the assembler take a -shared flag which means
approximately what the -buildmode=c-shared flag means in the proposed "Go
Execution Modes" document. As part of implementing other modes, the term
"shared" becomes horribly overloaded, so this replaces -shared with a
-buildmode argument instead (which currently only handles -buildmode=c-shared,
no new behaviour here).

As these tools did not take a -shared flag in 1.4 I've not retained support for
it.

Change-Id: I1c36f0060e0c3d256f36d2e9a07990446b756379