runtime: add ragged barrier function
DRAFT. This appears to work and I like the API, but I'm still feeling
my way around the scheduler (partly I did this as an exercise in
better understanding the P transitions).

This adds forEachP, which performs a general-purpose ragged GC
barrier. forEachP takes a callback and invokes it for every P at a GC
safe point. Ps that are idle or in a syscall are considered to be at a
continuous safe point.

To handle the case where a P wants to exit idle or a syscall while
forEachP is running the callback on that P's behalf, this introduces a
new _Pgcwait status that causes the exiting P to spin until this brief
race is resolved.

We'll use forEachP to replace the STW around enabling the write
barrier and to replace the current asynchronous per-M wbuf cache with
a cooperatively managed per-P gcWork cache.

Change-Id: Ie944f8ce1fead7c79bf271d2f42fcd61a41bb3cc