runtime: replace STW for enabling write barriers with ragged barrier
DRAFT (because forEachP is a draft).

Currently, we use a full stop-the-world around enabling write
barriers. This is to ensure that all Gs have enabled write barriers
before any marking occurs (either in func gc() or in gchelpwork()).

However, there's no need to bring the whole world to a synchronous
stop to ensure this. This change replaces the STW with a ragged
barrier that ensures each P has individually observed that write
barriers should be enabled before GC enables marking.

As part of this, we introduce a new _GCmarksetup phase that enables
write barriers but not mutator-assisted marking. This is necessary
because the existing _GCmark phase enables both, but write barriers
must be enabled strictly before mutator-assisted marking is enabled.

Change-Id: If2f129a6a55bd8bdd4308067af2b739f3fb41955