Update patch set 3
Patch Set 3:

> Out of curiosity, did you try this with tip or 1.4? Russ recently
 > added some limited transitive inlining, but I don't know whether it
 > covers this.
 > 
 > When you open the issue, please include your strings.go change or,
 > better, a standalone variant of it. It always helps to have
 > concrete code to look at. Thanks! And nice find.

I originally tried it with both 1.4.2 and Tip, but I uploaded the wrong benchmark results - I've corrected the CL to show the difference with the change relative to Tip. The recent changes removed one of the allocs that were being done, but there is still another from the looks of it.

Patch-set: 3