Update patch set 2
Patch Set 2:

(1 comment)

> Hmmm. Could non-escaping convT2[E|I] calls can be inlined? That'd
 > be faster and shorter, and we could dispense with the extra param.

The tendency is that we inline most of these conversions. I agree that it makes sense for the non-escaping case. But these changes are about different things: escape analysis vs codegen; and with different risk values. I would prefer it they go in as separate changes.

Patch-set: 2