runtime,reflect: support multiple moduledata objects
This changes all the places that consult themoduledata to consult a
linked list of moduledata objects, as will be necessary for
-buildmode=shered.

Obviously, as there is as yet no way of adding moduledata objects to
this list, all this change achieves right now is wasting a few
instructions here and there.  But I'd love to know if any of the GC
guys see anything that looks super scary in this.

Change-Id: I397af7f60d0849b76aaccedf72238fe664867051