reflect,cmd/internal/gc: look for pointer types by string before synthesizing
NOT FOR MERGING

When there are shared libraries in play one can't rely on the ptrto
field of the type data: a type T may be defined in a module that makes
no use of pointers to that type, but a module that imports the first
one certainly can use pointers to that type, and so ends up defining a
type data for the pointer type and type.*T symbol pointing at it.
It's important that calling .PtrTo() on the refect.Type for T returns
this type data and not some synthesized object.

Fortunately, the reflect package already has a mechanism for doing
this sort of thing: ChanOf/MapOf/etc look for pre-existing type data
by name.  So this change just extends PtrTo() to consult this too, and
changes the compiler to include pointer types in the data consulted.

Probably I should avoid adding pointers to the typelink data when not
building for shared, but that can be added easily after I've added the
buildmode flags to 6g etc.

Change-Id: I3773c066fd0679a62e9fc52a84bf64f1d67662b7